Upsample offset
