+++ crates/polars-time/src/upsample.rs
@@ -43,7 +43,6 @@ pub trait PolarsUpsample {
         by: I,
         time_column: &str,
         every: Duration,
-        offset: Duration,
     ) -> PolarsResult<DataFrame>;
 
     /// Upsample a [`DataFrame`] at a regular frequency.
@@ -85,7 +84,6 @@ pub trait PolarsUpsample {
         by: I,
         time_column: &str,
         every: Duration,
-        offset: Duration,
     ) -> PolarsResult<DataFrame>;
 }
 
@@ -95,13 +93,11 @@ impl PolarsUpsample for DataFrame {
         by: I,
         time_column: &str,
         every: Duration,
-        offset: Duration,
     ) -> PolarsResult<DataFrame> {
         let by = by.into_vec();
         let time_type = self.column(time_column)?.dtype();
-        ensure_duration_matches_data_type(offset, time_type, "offset")?;
         ensure_duration_matches_data_type(every, time_type, "every")?;
-        upsample_impl(self, by, time_column, every, offset, false)
+        upsample_impl(self, by, time_column, every, false)
     }
 
     fn upsample_stable<I: IntoVec<String>>(
@@ -109,13 +105,11 @@ impl PolarsUpsample for DataFrame {
         by: I,
         time_column: &str,
         every: Duration,
-        offset: Duration,
     ) -> PolarsResult<DataFrame> {
         let by = by.into_vec();
         let time_type = self.column(time_column)?.dtype();
-        ensure_duration_matches_data_type(offset, time_type, "offset")?;
         ensure_duration_matches_data_type(every, time_type, "every")?;
-        upsample_impl(self, by, time_column, every, offset, true)
+        upsample_impl(self, by, time_column, every, true)
     }
 }
 
@@ -124,7 +118,6 @@ fn upsample_impl(
     by: Vec<String>,
     index_column: &str,
     every: Duration,
-    offset: Duration,
     stable: bool,
 ) -> PolarsResult<DataFrame> {
     let s = source.column(index_column)?;
@@ -137,7 +130,7 @@ fn upsample_impl(
                 .unwrap()
         })
         .unwrap();
-        let mut out = upsample_impl(&df, by, index_column, every, offset, stable)?;
+        let mut out = upsample_impl(&df, by, index_column, every, stable)?;
         out.apply(index_column, |s| s.cast(time_type).unwrap())
             .unwrap();
         Ok(out)
@@ -154,7 +147,7 @@ fn upsample_impl(
                 .unwrap()
         })
         .unwrap();
-        let mut out = upsample_impl(&df, by, index_column, every, offset, stable)?;
+        let mut out = upsample_impl(&df, by, index_column, every, stable)?;
         out.apply(index_column, |s| s.cast(time_type).unwrap())
             .unwrap();
         Ok(out)
@@ -165,13 +158,13 @@ fn upsample_impl(
                 .unwrap()
         })
         .unwrap();
-        let mut out = upsample_impl(&df, by, index_column, every, offset, stable)?;
+        let mut out = upsample_impl(&df, by, index_column, every, stable)?;
         out.apply(index_column, |s| s.cast(time_type).unwrap())
             .unwrap();
         Ok(out)
     } else if by.is_empty() {
         let index_column = source.column(index_column)?;
-        upsample_single_impl(source, index_column, every, offset)
+        upsample_single_impl(source, index_column, every)
     } else {
         let gb = if stable {
             source.group_by_stable(by)
@@ -181,7 +174,7 @@ fn upsample_impl(
         // don't parallelize this, this may SO on large data.
         gb?.apply(|df| {
             let index_column = df.column(index_column)?;
-            upsample_single_impl(&df, index_column, every, offset)
+            upsample_single_impl(&df, index_column, every)
         })
     }
 }
@@ -190,7 +183,6 @@ fn upsample_single_impl(
     source: &DataFrame,
     index_column: &Series,
     every: Duration,
-    offset: Duration,
 ) -> PolarsResult<DataFrame> {
     let index_col_name = index_column.name();
 
@@ -207,11 +199,6 @@ fn upsample_single_impl(
                         #[cfg(feature = "timezones")]
                         Some(tz) => Some(parse_time_zone(tz)?),
                         _ => None,
-                    };
-                    let first = match tu {
-                        TimeUnit::Nanoseconds => offset.add_ns(first, tz.as_ref())?,
-                        TimeUnit::Microseconds => offset.add_us(first, tz.as_ref())?,
-                        TimeUnit::Milliseconds => offset.add_ms(first, tz.as_ref())?,
                     };
                     let range = datetime_range_impl(
                         index_col_name,
