+++ py-polars/src/dataframe/general.rs
@@ -590,23 +590,13 @@ impl PyDataFrame {
         by: Vec<String>,
         index_column: &str,
         every: &str,
-        offset: &str,
         stable: bool,
     ) -> PyResult<Self> {
         let out = if stable {
-            self.df.upsample_stable(
-                by,
-                index_column,
-                Duration::parse(every),
-                Duration::parse(offset),
-            )
+            self.df
+                .upsample_stable(by, index_column, Duration::parse(every))
         } else {
-            self.df.upsample(
-                by,
-                index_column,
-                Duration::parse(every),
-                Duration::parse(offset),
-            )
+            self.df.upsample(by, index_column, Duration::parse(every))
         };
         let out = out.map_err(PyPolarsErr::from)?;
         Ok(out.into())
